requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiohttp>=3.11.11
click<=7.1.2
cryptography>=44.0.0
halo>=0.0.31
Jinja2>=3.1.5
netmiko>=4.5.0
pydantic>=2.10.5
Pygments>=2.19.1
PyYAML>=6.0.2
requests>=2.32.3
rich>=13.9.4
tablib>=3.7.0
tabulate>=0.9.0
typer==0.10.0
urllib3>=2.3.0
yarl>=1.18.3
